+++ src/ATrous/atrous_3d_reconstruct.cc
@@ -43,7 +43,7 @@ using std::setw;
 namespace duchamp
 {
 
-  void atrous3DReconstruct(long &xdim, long &ydim, long &zdim, float *&input, 
+  void atrous3DReconstruct(unsigned long &xdim, unsigned long &ydim, unsigned long &zdim, float *&input, 
 			   float *&output, Param &par)
   {
     ///  A routine that uses the a trous wavelet method to reconstruct a 
@@ -62,24 +62,24 @@ namespace duchamp
     ///  \param output The returned reconstructed spectrum. This array needs to be declared beforehand.
     ///  \param par The Param set.
 
-    long size = xdim * ydim * zdim;
-    long spatialSize = xdim * ydim;
-    long mindim = xdim;
+    size_t size = xdim * ydim * zdim;
+    size_t spatialSize = xdim * ydim;
+    unsigned long mindim = xdim;
     if (ydim<mindim) mindim = ydim;
     if (zdim<mindim) mindim = zdim;
-    int numScales = par.filter().getNumScales(mindim);
+    unsigned int numScales = par.filter().getNumScales(mindim);
 
     double *sigmaFactors = new double[numScales+1];
-    for(int i=0;i<=numScales;i++){
-      if(i<=par.filter().maxFactor(3)) 
+    for(size_t i=0;i<=numScales;i++){
+      if(i<=size_t(par.filter().maxFactor(3)) )
 	sigmaFactors[i] = par.filter().sigmaFactor(3,i);
       else sigmaFactors[i] = sigmaFactors[i-1] / sqrt(8.);
     }
 
     float mean,sigma,originalSigma,originalMean,oldsigma,newsigma;
     bool *isGood = new bool[size];
-    int goodSize=0;
-    for(int pos=0;pos<size;pos++){
+    size_t goodSize=0;
+    for(size_t pos=0;pos<size;pos++){
       isGood[pos] = !par.isBlank(input[pos]);
       if(isGood[pos]) goodSize++;
     }
@@ -88,7 +88,7 @@ namespace duchamp
       // There are no good pixels -- everything is BLANK for some reason.
       // Return the input array as the output, and give a warning message.
 
-      for(int pos=0;pos<xdim; pos++) output[pos] = input[pos];
+      for(size_t pos=0;pos<xdim; pos++) output[pos] = input[pos];
 
       duchampWarning("3D Reconstruction",
 		     "There are no good pixels to be reconstructed -- all are BLANK.\nPerhaps you need to try this with flagTrim=false.\nReturning input array.\n");
@@ -104,16 +104,16 @@ namespace duchamp
       float *wavelet = new float[size];
       float *residual = new float[size];
 
-      for(int pos=0;pos<size;pos++) output[pos]=0.;
+      for(size_t pos=0;pos<size;pos++) output[pos]=0.;
 
       // Define the 3-D (separable) filter, using info from par.filter()
-      int filterwidth = par.filter().width();
+      size_t filterwidth = par.filter().width();
       int filterHW = filterwidth/2;
-      int fsize = filterwidth*filterwidth*filterwidth;
+      size_t fsize = filterwidth*filterwidth*filterwidth;
       double *filter = new double[fsize];
-      for(int i=0;i<filterwidth;i++){
-	for(int j=0;j<filterwidth;j++){
-	  for(int k=0;k<filterwidth;k++){
+      for(size_t i=0;i<filterwidth;i++){
+	for(size_t j=0;j<filterwidth;j++){
+	  for(size_t k=0;k<filterwidth;k++){
 	    filter[i +j*filterwidth + k*filterwidth*filterwidth] = 
 	      par.filter().coeff(i) * par.filter().coeff(j) * par.filter().coeff(k);
 	  }
@@ -124,20 +124,20 @@ namespace duchamp
       //  Only do this if flagBlankPix is true. 
       //  Otherwise use the full range of x and y.
       //  No trimming is done in the z-direction at this point.
-      int *xLim1 = new int[ydim];
-      for(int i=0;i<ydim;i++) xLim1[i] = 0;
-      int *xLim2 = new int[ydim];
-      for(int i=0;i<ydim;i++) xLim2[i] = xdim-1;
-      int *yLim1 = new int[xdim];
-      for(int i=0;i<xdim;i++) yLim1[i] = 0;
-      int *yLim2 = new int[xdim];
-      for(int i=0;i<xdim;i++) yLim2[i] = ydim-1;
+      long *xLim1 = new long[ydim];
+      for(size_t i=0;i<ydim;i++) xLim1[i] = 0;
+      long *xLim2 = new long[ydim];
+      for(size_t i=0;i<ydim;i++) xLim2[i] = xdim-1;
+      long *yLim1 = new long[xdim];
+      for(size_t i=0;i<xdim;i++) yLim1[i] = 0;
+      long *yLim2 = new long[xdim];
+      for(size_t i=0;i<xdim;i++) yLim2[i] = ydim-1;
 
       if(par.getFlagBlankPix()){
 	float avGapX = 0, avGapY = 0;
-	for(int row=0;row<ydim;row++){
-	  int ct1 = 0;
-	  int ct2 = xdim - 1;
+	for(size_t row=0;row<ydim;row++){
+	  size_t ct1 = 0;
+	  size_t ct2 = xdim - 1;
 	  while((ct1<ct2)&&(par.isBlank(input[row*xdim+ct1]))) ct1++;
 	  while((ct2>ct1)&&(par.isBlank(input[row*xdim+ct2]))) ct2--;
 	  xLim1[row] = ct1;
@@ -146,9 +146,9 @@ namespace duchamp
 	}
 	avGapX /= float(ydim);
 
-	for(int col=0;col<xdim;col++){
-	  int ct1=0;
-	  int ct2=ydim-1;
+	for(size_t col=0;col<xdim;col++){
+	  size_t ct1=0;
+	  size_t ct2=ydim-1;
 	  while((ct1<ct2)&&(par.isBlank(input[col+xdim*ct1]))) ct1++;
 	  while((ct2>ct1)&&(par.isBlank(input[col+xdim*ct2]))) ct2--;
 	  yLim1[col] = ct1;
@@ -165,16 +165,16 @@ namespace duchamp
       float threshold;
       int iteration=0;
       newsigma = 1.e9;
-      for(int i=0;i<size;i++) output[i] = 0;
+      for(size_t i=0;i<size;i++) output[i] = 0;
       do{
 	if(par.isVerbose()) std::cout << "Iteration #"<<setw(2)<<++iteration<<": ";
 	// first, get the value of oldsigma, set it to the previous newsigma value
 	oldsigma = newsigma;
 	// we are transforming the residual array (input array first time around)
-	for(int i=0;i<size;i++)  coeffs[i] = input[i] - output[i];
+	for(size_t i=0;i<size;i++)  coeffs[i] = input[i] - output[i];
 
 	int spacing = 1;
-	for(int scale = 1; scale<=numScales; scale++){
+	for(unsigned int scale = 1; scale<=numScales; scale++){
 
 	  if(par.isVerbose()){
 	    std::cout << "Scale ";
@@ -183,10 +183,10 @@ namespace duchamp
 	    std::cout << std::flush;
 	  }
 
-	  int pos = -1;
-	  for(int zpos = 0; zpos<zdim; zpos++){
-	    for(int ypos = 0; ypos<ydim; ypos++){
-	      for(int xpos = 0; xpos<xdim; xpos++){
+	  size_t pos = -1;
+	  for(unsigned long zpos = 0; zpos<zdim; zpos++){
+	    for(unsigned long ypos = 0; ypos<ydim; ypos++){
+	      for(unsigned long xpos = 0; xpos<xdim; xpos++){
 		// loops over each pixel in the image
 		pos++;
 
@@ -195,30 +195,31 @@ namespace duchamp
 		if(!isGood[pos] )  wavelet[pos] = 0.;
 		else{
 
-		  int filterpos = -1;
+		  unsigned int filterpos = -1;
 		  for(int zoffset=-filterHW; zoffset<=filterHW; zoffset++){
-		    int z = zpos + spacing*zoffset;
+		    long z = zpos + spacing*zoffset;
 		    if(z<0) z = -z;                 // boundary conditions are 
-		    if(z>=zdim) z = 2*(zdim-1) - z; //    reflection.
+		    if(z>=long(zdim)) z = 2*(zdim-1) - z; //    reflection.
 		
-		    int oldchan = z * spatialSize;
+		    size_t oldchan = z * spatialSize;
 		
 		    for(int yoffset=-filterHW; yoffset<=filterHW; yoffset++){
-		      int y = ypos + spacing*yoffset;
+		      long y = long(ypos) + spacing*yoffset;
 
 		      // Boundary conditions -- assume reflection at boundaries.
 		      // Use limits as calculated above
 		      if(yLim1[xpos]!=yLim2[xpos]){ 
 			// if these are equal we will get into an infinite loop
 			while((y<yLim1[xpos])||(y>yLim2[xpos])){
+			// std::cerr << y << " " <<spacing << " " << yoffset << " " << ypos << " " << xpos << " " << yLim1[xpos] << " " << yLim2[xpos] << "\n";
 			  if(y<yLim1[xpos]) y = 2*yLim1[xpos] - y;      
 			  else if(y>yLim2[xpos]) y = 2*yLim2[xpos] - y;      
 			}
 		      }
-		      int oldrow = y * xdim;
+		      size_t oldrow = y * xdim;
 	  
 		      for(int xoffset=-filterHW; xoffset<=filterHW; xoffset++){
-			int x = xpos + spacing*xoffset;
+			long x = long(xpos) + spacing*xoffset;
 
 			// Boundary conditions -- assume reflection at boundaries.
 			// Use limits as calculated above
@@ -230,7 +231,7 @@ namespace duchamp
 			  }
 			}
 
-			int oldpos = oldchan + oldrow + x;
+			size_t oldpos = oldchan + oldrow + x;
 
 			filterpos++;
 		   
@@ -247,7 +248,7 @@ namespace duchamp
 	  } //-> end of zpos loop
 
 	  // Need to do this after we've done *all* the convolving
-	  for(int pos=0;pos<size;pos++) coeffs[pos] = coeffs[pos] - wavelet[pos];
+	  for(size_t pos=0;pos<size;pos++) coeffs[pos] = coeffs[pos] - wavelet[pos];
 
 	  // Have found wavelet coeffs for this scale -- now threshold
 	  if(scale>=par.getMinScale()){
@@ -255,7 +256,7 @@ namespace duchamp
 	
 	    threshold = mean + 
 	      par.getAtrousCut()*originalSigma*sigmaFactors[scale];
-	    for(int pos=0;pos<size;pos++){
+	    for(size_t pos=0;pos<size;pos++){
 	      if(!isGood[pos]){
 		output[pos] = input[pos]; 
 		// this preserves the Blank pixel values in the output.
@@ -271,9 +272,9 @@ namespace duchamp
 
 	} //-> end of scale loop 
 
-	for(int pos=0;pos<size;pos++) if(isGood[pos]) output[pos] += coeffs[pos];
+	for(size_t pos=0;pos<size;pos++) if(isGood[pos]) output[pos] += coeffs[pos];
 
-	for(int i=0;i<size;i++) residual[i] = input[i] - output[i];
+	for(size_t i=0;i<size;i++) residual[i] = input[i] - output[i];
 	findMedianStats(residual,goodSize,isGood,mean,newsigma);
 	newsigma = madfmToSigma(newsigma);
